requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
channels==4.0.0
Django==4.2.4
django-filter==23.5
django_multiselectfield==0.1.12
django_notifications_hq==1.8.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1